requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.1.13
django-ckeditor==6.0.0
django-environ==0.11.2
django-js-asset==1.2.2
filelock==3.15.4
fsspec==2024.6.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
promise==2.3
py4j==0.10.9.7
pyjokes==0.6.0
pyspark==3.5.2
pytz==2021.1
requests==2.32.3
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
text-unidecode==1.3
torch==2.4.0
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
